refactor: refactored auth controller
:recycle:
refactored the auth controller package and moved DTO and DRO Object to their own packages
