+++ src/main/java/com/quiz_app/QuizApplication.java
@@ -4,7 +4,7 @@ import com.quiz_app.entity.quiz.Question;
 import com.quiz_app.entity.quiz.Quiz;
 import com.quiz_app.repository.QuizRepository;
 import com.quiz_app.service.auth.AuthenticationService;
-import com.quiz_app.controller.authcontroller.RegisterRequest;
+import com.quiz_app.controller.authcontroller.request.RegisterRequest;
 import org.springframework.boot.CommandLineRunner;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
