+++ src/main/java/com/quiz_app/controller/authcontroller/request/ForgotPasswordVerificationRequest.java
@@ -1,4 +1,4 @@
-package com.quiz_app.controller.authcontroller;
+package com.quiz_app.controller.authcontroller.request;
 
 import lombok.Builder;
 import lombok.Data;
