+++ src/main/java/com/quiz_app/controller/authcontroller/request/RegisterRequest.java
@@ -1,4 +1,4 @@
-package com.quiz_app.controller.authcontroller;
+package com.quiz_app.controller.authcontroller.request;
 
 import com.quiz_app.entity.user.Role;
 import jakarta.validation.constraints.Email;
