+++ src/main/java/com/quiz_app/controller/quizresourcecontroller/authenticatedquizcontroller/AuthenticatedQuizController.java
@@ -1,0 +1,13 @@
+package com.quiz_app.controller.quizresourcecontroller.authenticatedquizcontroller;
+
+import lombok.RequiredArgsConstructor;
+import org.springframework.web.bind.annotation.CrossOrigin;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+@RequestMapping("/api/v1/quiz/authenticated/")
+@RequiredArgsConstructor
+@CrossOrigin(origins = "http://localhost:3000", maxAge = 3600)
+public class AuthenticatedQuizController {
+}
